sync-github-releases: give the tag scheme its own module
getTagScheme() and the TagScheme type are an independent concept — how a
package names and recognizes its release tags — and are now consumed by
syncPackage() as well as release-plan.ts. Move them (with their tests) out of
release-plan.ts into sync/tag-scheme.ts, so each file is one abstraction:
tag-scheme.ts owns tag naming/ownership, release-plan.ts owns reconciliation.
No behavior change.
