+++ .github/workflows/sync-github-releases/sync.ts
@@ -6,7 +6,8 @@ main()
 import { readFile } from 'node:fs/promises'
 import path from 'node:path'
 import { applyReleasePlan } from './sync/apply-release-plan.ts'
-import { getReleasePlan, getTagScheme } from './sync/release-plan.ts'
+import { getReleasePlan } from './sync/release-plan.ts'
+import { getTagScheme } from './sync/tag-scheme.ts'
 import { getReleaseNotesByVersion, type ReleaseNotesByVersion } from './utils/changelog.ts'
 import { getPushedFiles, getRepoRoot, getTrackedChangelogFiles, toPackageDirs } from './utils/git.ts'
 import {
