+++ .github/workflows/sync-github-releases/sync/release-plan.ts
@@ -1,10 +1,10 @@
 export { getReleasePlan }
-export { getTagScheme }
 export type { ReleasePlan }
 export type { ReleaseToCreate }
 
 import type { ReleaseNotesByVersion } from '../utils/changelog.ts'
 import type { Release } from '../utils/github.ts'
+import type { TagScheme } from './tag-scheme.ts'
 
 type ReleasePlan = {
   releasesToCreate: ReleaseToCreate[]
@@ -27,32 +27,6 @@ type ReleaseToUpdate = {
 type ReleaseToDelete = {
   release_id: number
   tag_name: string
-}
-
-// A package's release-tag scheme, defined in one place so building a tag and recognizing one can't
-// drift apart. `build` turns a version into its tag; `owns` reports whether a release tag is this
-// package's — i.e. a candidate for deletion once its changelog entry is gone, never another package's
-// release nor a tag we don't create (e.g. `nightly`).
-//
-// A single package keeps the historical bare `vX.Y.Z` tag. Several packages share the repo's tag
-// namespace, so their tags are qualified with the package name (e.g. `create-vike-core@0.0.391`) to
-// avoid collisions.
-type TagScheme = {
-  build(version: string): string
-  owns(releaseTag: string): boolean
-}
-function getTagScheme(packageName: string, hasMultiplePackages: boolean): TagScheme {
-  if (hasMultiplePackages) {
-    const prefix = `${packageName}@`
-    return {
-      build: (version) => `${prefix}${version}`,
-      owns: (releaseTag) => releaseTag.startsWith(prefix),
-    }
-  }
-  return {
-    build: (version) => `v${version}`,
-    owns: (releaseTag) => /^v\d+\.\d+\.\d+/.test(releaseTag),
-  }
 }
 
 function getReleasePlan({
